Refactor native_functions.py
