+++ tools/codegen/dest/__init__.py
@@ -1,6 +1,6 @@
 from .lazy_ir import LazyIR as LazyIR
 from .ts_lowering import TsLowering as TsLowering
+from .lazy_ir import compute_lazy_shape_dtype_decl as compute_lazy_shape_dtype_decl
+from .lazy_ir import compute_lazy_native_function_definition as compute_lazy_native_function_definition
 from .register_dispatch_key import RegisterDispatchKey as RegisterDispatchKey
-from .native_functions import compute_lazy_shape_dtype_decl as compute_lazy_shape_dtype_decl
 from .native_functions import compute_native_function_declaration as compute_native_function_declaration
-from .native_functions import compute_lazy_native_function_definition as compute_lazy_native_function_definition
